Improved network version exchange protocol
This commit relaxes the check on the client side to allow
connecting to servers with a more recent network protocol in the
future. Clients also send their protocol version to the server.
This information is simply ignored by older servers and does not
break compatibility.

The idea is to create a transition period after which the protocol
can be modified (to exchange more information at connection time,
such as support for stereo streams) without breaking compatibilty
for clients and servers based on deflect >= 0.12.1.